chore: Extract ReflectionUtil#resolveTypeParameter from TypeParameterMatcher.
